# src/core/Snakk.Sdk/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Snakk.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:17100/"
    }
  ],
  "paths": {
...
    "/hubs/{hubId}/spaces": {
      "get": {
        "tags": [
          "Hubs"
        ],
        "operationId": "GetSpacesByHub",
        "parameters": [
          {
            "name": "hubId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseOfSpaceByHubResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/spaces": {
      "post": {
        "tags": [
          "Spaces"
        ],
        "operationId": "CreateSpace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/spaces/{publicId}": {
      "get": {
        "tags": [
          "Spaces"
        ],
        "operationId": "GetSpace",
        "parameters": [
          {
            "name": "publicId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/spaces/by-slug/{slug}": {
      "get": {
        "tags": [
          "Spaces"
        ],
        "operationId": "GetSpaceBySlug",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/spaces/{spaceId}/discussions": {
      "get": {
        "tags": [
          "Spaces"
        ],
        "operationId": "GetDiscussionsBySpace",
        "parameters": [
          {
            "name": "spaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseOfDiscussionBySpaceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/spaces/top-active-today": {
      "get": {
        "tags": [
          "Spaces"
        ],
        "operationId": "GetTopActiveSpacesToday",
        "parameters": [
          {
            "name": "hubId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "communityId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopActiveSpacesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/spaces/{spaceId}/rules": {
      "get": {
        "tags": [
          "Spaces"
        ],
        "operationId": "GetSpaceRules",
        "parameters": [
          {
            "name": "spaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceRulesDto"
                }
              }
            }
          }
        }
      }
    },
    "/spaces/{publicId}/stats": {
      "get": {
        "tags": [
          "Spaces"
        ],
        "operationId": "GetSpaceStats",
        "parameters": [
          {
            "name": "publicId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceStatsResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/content/spaces": {
      "get": {
        "tags": [
          "Admin - Content"
        ],
        "operationId": "AdminGetSpaces",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hubId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSpaceListResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/c/{slug}/manage/spaces": {
      "get": {
        "tags": [
          "Community Management"
        ],
        "operationId": "GetCommunitySpaces",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/c/{communitySlug}/h/{hubSlug}/manage/spaces": {
      "get": {
        "tags": [
          "Hub Management"
        ],
        "operationId": "GetHubSpaces",
        "parameters": [
          {
            "name": "communitySlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hubSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/avatars/space/{publicId}": {
      "get": {
        "tags": [
          "Avatars"
        ],
        "operationId": "GetSpaceAvatar",
        "parameters": [
          {
            "name": "publicId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminSpaceItemResponse": {
        "required": [
          "id",
          "name",
          "slug",
          "description",
          "hubId",
          "hubName",
          "communityId",
          "communityName",
          "createdAt",
          "discussionCount"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "hubId": {
            "type": "string"
          },
          "hubName": {
            "type": "string"
          },
          "communityId": {
            "type": "string"
          },
          "communityName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "discussionCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "AdminSpaceListResponse": {
        "required": [
          "spaces",
          "total",
          "page",
          "pageSize"
        ],
        "type": "object",
        "properties": {
          "spaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminSpaceItemResponse"
            }
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "AuthorRef": {
        "required": [
          "publicId",
          "displayName"
        ],
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "avatarUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "role": {
            "type": [
              "null",
              "string"
            ]
          },
          "isDeleted": {
            "type": "boolean",
            "default": false
          }
        }
      },
...
      "CreateSpaceRequest": {
        "required": [
          "hubId",
          "name",
          "slug",
          "description"
        ],
        "type": "object",
        "properties": {
          "hubId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "DiscussionBySpaceResponse": {
        "required": [
          "publicId",
          "spaceId",
          "title",
          "slug",
          "createdAt",
          "lastActivityAt",
          "isPinned",
          "isLocked",
          "postCount",
          "reactionCount",
          "author",
          "tags"
        ],
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string"
          },
          "spaceId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastActivityAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "isPinned": {
            "type": "boolean"
          },
          "isLocked": {
            "type": "boolean"
          },
          "postCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "reactionCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "author": {
            "$ref": "#/components/schemas/AuthorRef"
          },
          "tags": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "EntityRef": {
        "required": [
          "publicId",
          "slug",
          "name"
        ],
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "avatarUrl": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "LatestDiscussionRef": {
        "required": [
          "publicId",
          "title",
          "slug",
          "lastActivityAt",
          "authorPublicId",
          "authorDisplayName",
          "authorAvatarFileName",
          "postCount"
        ],
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "lastActivityAt": {
            "type": "string",
            "format": "date-time"
          },
          "authorPublicId": {
            "type": "string"
          },
          "authorDisplayName": {
            "type": "string"
          },
          "authorAvatarFileName": {
            "type": [
              "null",
              "string"
            ]
          },
          "postCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "PagedResponseOfDiscussionBySpaceResponse": {
        "required": [
          "items",
          "offset",
          "pageSize",
          "hasMoreItems"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiscussionBySpaceResponse"
            }
          },
          "offset": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "hasMoreItems": {
            "type": "boolean"
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "PagedResponseOfSpaceByHubResponse": {
        "required": [
          "items",
          "offset",
          "pageSize",
          "hasMoreItems"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpaceByHubResponse"
            }
          },
          "offset": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "hasMoreItems": {
            "type": "boolean"
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SpaceByHubResponse": {
        "required": [
          "publicId",
          "hubPublicId",
          "name",
          "slug",
          "description",
          "createdAt",
          "discussionCount",
          "replyCount"
        ],
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string"
          },
          "hubPublicId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "discussionCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "replyCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "latestDiscussion": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LatestDiscussionRef"
              }
            ]
          }
        }
      },
...
      "SpaceResponse": {
        "required": [
          "publicId",
          "hubId",
          "name",
          "slug",
          "description",
          "createdAt",
          "discussionCount",
          "replyCount"
        ],
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string"
          },
          "hubId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "discussionCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "replyCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "lastModifiedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "latestDiscussion": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LatestDiscussionRef"
              }
            ]
          }
        }
      },
      "SpaceRuleDto": {
        "required": [
          "title",
          "description"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "title": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "description": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string"
          },
          "order": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "SpaceRulesDto": {
        "type": "object",
        "properties": {
          "rules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpaceRuleDto"
            }
          }
        }
      },
      "SpaceStatsResponse": {
        "required": [
          "publicId",
          "name",
          "avatarUrl",
          "discussionCount",
          "replyCount",
          "followerCount"
        ],
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "avatarUrl": {
            "type": "string"
          },
          "discussionCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "replyCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "followerCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "TopActiveSpaceResponse": {
        "required": [
          "publicId",
          "name",
          "slug",
          "postCountToday",
          "hub"
        ],
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "postCountToday": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "hub": {
            "$ref": "#/components/schemas/EntityRef"
          }
        }
      },
      "TopActiveSpacesResponse": {
        "required": [
          "items"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TopActiveSpaceResponse"
            }
          }
        }
      },
...
    }
  },
...
}