# src/services/Snakk.Api/Endpoints/SpaceEndpoints.cs
namespace Snakk.Api.Endpoints;

using Snakk.Api.Models;
using Snakk.Application.DTOs.Management;
using Snakk.Application.DTOs.Responses;
using Snakk.Application.DTOs.Stats;
using Snakk.Application.Services;
using Snakk.Application.UseCases;
using Snakk.Domain.ValueObjects;
using Snakk.Shared.Helpers;

public static class SpaceEndpoints
{
    public static void MapSpaceEndpoints(this IEndpointRouteBuilder app)
    {
        var group = app.MapGroup("/spaces")
            .WithTags("Spaces");

        group.MapPost("/", CreateSpaceAsync)
            .WithName("CreateSpace")
            .Produces<SpaceResponse>(StatusCodes.Status201Created);

        group.MapGet("/{publicId}", GetSpaceAsync)
            .WithName("GetSpace")
            .Produces<SpaceResponse>();

        group.MapGet("/by-slug/{slug}", GetSpaceBySlugAsync)
            .WithName("GetSpaceBySlug")
            .Produces<SpaceResponse>();

        group.MapGet("/{spaceId}/discussions", GetDiscussionsBySpaceAsync)
            .WithName("GetDiscussionsBySpace")
            .Produces<PagedResponse<DiscussionBySpaceResponse>>();

        group.MapGet("/top-active-today", GetTopActiveSpacesTodayAsync)
            .WithName("GetTopActiveSpacesToday")
            .Produces<TopActiveSpacesResponse>();

        group.MapGet("/{spaceId}/rules", GetSpaceRulesAsync)
            .WithName("GetSpaceRules")
            .Produces<SpaceRulesDto>();

        group.MapGet("/{publicId}/stats", GetSpaceStatsAsync)
            .WithName("GetSpaceStats")
            .Produces<SpaceStatsResponse>();
    }

    private static async Task<IResult> CreateSpaceAsync(
        CreateSpaceRequest request,
        SpaceUseCase useCase)
    {
        var result = await useCase.CreateSpaceAsync(
            HubId.From(request.HubId),
            request.Name,
            request.Slug,
            request.Description);

        if (!result.IsSuccess)
            return Results.BadRequest(new { error = result.Error });

        return TypedResults.Created($"/spaces/{result.Value!.PublicId}", new SpaceResponse(
            PublicId: result.Value.PublicId.Value,
            HubId: result.Value.HubId.Value,
            Name: result.Value.Name,
            Slug: result.Value.Slug,
            Description: result.Value.Description,
            CreatedAt: result.Value.CreatedAt,
            DiscussionCount: 0,
            ReplyCount: 0));
    }

    private static async Task<IResult> GetSpaceAsync(
        string publicId,
        SpaceUseCase useCase)
    {
        var result = await useCase.GetSpaceAsync(SpaceId.From(publicId));

        if (!result.IsSuccess)
            return Results.NotFound(new { error = result.Error });

        return TypedResults.Ok(new SpaceResponse(
            PublicId: result.Value!.PublicId.Value,
            HubId: result.Value.HubId.Value,
            Name: result.Value.Name,
            Slug: result.Value.Slug,
            Description: result.Value.Description,
            CreatedAt: result.Value.CreatedAt,
            DiscussionCount: 0,
            ReplyCount: 0,
            LastModifiedAt: result.Value.LastModifiedAt));
    }

    private static async Task<IResult> GetSpaceBySlugAsync(
        string slug,
        SpaceUseCase useCase)
    {
        var result = await useCase.GetSpaceBySlugAsync(slug);

        if (!result.IsSuccess)
            return Results.NotFound(new { error = result.Error });

        return TypedResults.Ok(new SpaceResponse(
            PublicId: result.Value!.PublicId.Value,
            HubId: result.Value.HubId.Value,
            Name: result.Value.Name,
            Slug: result.Value.Slug,
            Description: result.Value.Description,
            CreatedAt: result.Value.CreatedAt,
            DiscussionCount: 0,
            ReplyCount: 0,
            LastModifiedAt: result.Value.LastModifiedAt));
    }

    private static async Task<IResult> GetDiscussionsBySpaceAsync(
        string spaceId,
        int offset,
        int pageSize,
        Application.Repositories.ISearchRepository searchRepo)
    {
        var result = await searchRepo.GetDiscussionsBySpaceAsync(spaceId, offset, pageSize);

        var items = result.Items.Select(d => new DiscussionBySpaceResponse(
            PublicId: d.PublicId,
            SpaceId: d.SpacePublicId,
            Title: d.Title,
            Slug: d.Slug,
            CreatedAt: d.CreatedAt,
            LastActivityAt: d.LastActivityAt ?? d.CreatedAt,
            IsPinned: d.IsPinned,
            IsLocked: d.IsLocked,
            PostCount: d.PostCount,
            ReactionCount: d.ReactionCount,
            Author: new AuthorRef(
                PublicId: d.AuthorPublicId,
                DisplayName: d.AuthorDisplayName,
                AvatarUrl: d.AuthorAvatarFileName),
            Tags: d.Tags?.Split(',', StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries)));

        return TypedResults.Ok(new PagedResponse<DiscussionBySpaceResponse>(
            Items: items,
            Offset: result.Offset,
            PageSize: result.PageSize,
            HasMoreItems: result.HasMoreItems));
    }

    private static async Task<IResult> GetTopActiveSpacesTodayAsync(
        StatisticsUseCase useCase,
        string? hubId = null,
        string? communityId = null)
    {
        var topSpaces = await useCase.GetTopActiveSpacesTodayAsync(hubId, communityId);

        var items = topSpaces.Select(s => new TopActiveSpaceResponse(
            PublicId: s.PublicId,
            Name: s.Name,
            Slug: s.Slug,
            PostCountToday: s.PostCountToday,
            Hub: new EntityRef(
                PublicId: s.HubPublicId,
                Slug: s.HubSlug,
                Name: s.HubName)));

        return TypedResults.Ok(new TopActiveSpacesResponse(items));
    }

    private static async Task<IResult> GetSpaceRulesAsync(
        string spaceId,
        ISpaceManagementService service,
        CancellationToken cancellationToken)
    {
        var rules = await service.GetRulesAsync(spaceId, cancellationToken);
        return Results.Ok(rules);
    }

    private static async Task<IResult> GetSpaceStatsAsync(string publicId, StatisticsUseCase useCase)
    {
        var result = await useCase.GetSpaceStatsAsync(publicId);

        if (!result.IsSuccess)
            return Results.NotFound();

        var stats = result.Value!;
        return TypedResults.Ok(new SpaceStatsResponse
        {
            PublicId = stats.PublicId,
            Name = stats.Name,
            Description = stats.Description,
            AvatarUrl = AvatarHelper.GetAvatarUrl(stats.PublicId, AvatarEntityType.Space, 0),
            DiscussionCount = stats.DiscussionCount,
            ReplyCount = stats.ReplyCount,
            FollowerCount = stats.FollowerCount
        });
    }
}